Raise an exception if `FLOW.read` violates the spec
The spec for FLOW.read says:

> The returned buffer will be of a size convenient to the flow
> implementation, but will always have at least 1 byte.

An implementation which returns a 0 length buffer is horribly buggy --
probably what's happening is the implementation has called a function
like read(2) and misinterpreted the return value. Note that, contrary
to the comment, the function `get_ibuf` which calls `ibuf_refill`
is actually "robust" in this case in that it will keep calling `FLOW.read`
in a loop, causing the program to spin forever.

The spec for the Mirage API says in general that:

> errors which represent programming errors such as assertion failures
> or illegal arguments are encoded as exceptions. The application may
> attempt to catch exceptions and recover or simply let the exception
> propagate and crash the application. If the application crashes then
> the runtime system should output diagnostics and abort.

Therefore we catch this programming error, emit a log message and throw
an exception hoping to crash the app immediately. It's likely the bug
will be very easily fixed once the error is triggered.

Signed-off-by: David Scott <dave@recoil.org>