Add an optional size limit to `read_line`
Previously a call to `read_line` would read and buffer arbitrary amounts of
data. If an untrusted entity supplies the data (e.g. consider a web-server reading
a request or header line) then the `read_line` can be provoked into allocating
so much memory that it kills the whole process / server.

This patch changes the API by adding an optional `?len` parameter (in the same
style as `read_some` and `read_exactly`) to bound the maximum length of the line.
If the bound is exceeded then the function returns an error (`Line_too_long`)

Users of this library dealing with untrusted data should decide on a suitable
`len` value to avoid possible memory exhaustion.

This patch includes a couple of extra unit tests.

Signed-off-by: David Scott <dave@recoil.org>