If the `read_line` limit is set too low, put the data back
Previously `read_line` would drop the data from the channel when
returning the `Line_too_long` error. This patch puts the data back
in the input queue to make it easier to handle this error. For
example an application would be able to call `read_some` or
`read_exactly` to peek at the data and make a better error message.

Signed-off-by: David Scott <dave@recoil.org>